--- code_snippets/java/src/main/java/develop/ContextKeyedExample.java
@@ -1,16 +1,0 @@
-package develop;
-
-import dev.restate.sdk.ObjectContext;
-
-import static example.generated.ExampleProto.*;
-import example.generated.ExampleRestate;
-
-public class ContextKeyedExample extends ExampleRestate.ExampleRestateImplBase {
-
-    // <start_here>
-    @Override
-    public void greet(ObjectContext ctx, ExampleRequest request) {
-        // ... Do something ... //
-    }
-    // <end_here>
-}
