+++ docs/develop/sdk/restate-context.mdx
@@ -48,13 +48,13 @@ The Restate context is supplied as the first argument `ctx` to the method.
 For unkeyed services:
 
 ```java
-CODE_LOAD::java/src/main/java/develop/ContextUnkeyedExample.java
+CODE_LOAD::java/src/main/java/develop/RestateContext.java#<start_unkeyed>-<end_unkeyed>
 ```
 
 For keyed services:
 
 ```java
-CODE_LOAD::java/src/main/java/develop/ContextKeyedExample.java
+CODE_LOAD::java/src/main/java/develop/RestateContext.java#<start_keyed>-<end_keyed>
 ```
 
 :::note
@@ -69,13 +69,13 @@ If you do not use Restate's generated protobuf code, you can access the restate 
 For unkeyed services:
 
 ```java
-CODE_LOAD::java/src/main/java/develop/ContextVanillaUnkeyedExample.java
+CODE_LOAD::java/src/main/java/develop/RestateContext.java#<start_grpc_unkeyed>-<end_grpc_unkeyed>
 ```
 
 For keyed services:
 
 ```java
-CODE_LOAD::java/src/main/java/develop/ContextVanillaKeyedExample.java
+CODE_LOAD::java/src/main/java/develop/RestateContext.java#<start_grpc_keyed>-<end_grpc_keyed>
 ```
 
 </details>
